Add vanilla css support to ContextMenu (#520)
* Add and update hero demos

* Consistency adjustment
